placing order using backend implemented
While posting the order to backend, I keep getting error 500.
First, I thought it is because wrong formatting, but it is not.
When formatting is incorrect, you get error 400.
I couldn't find the cause. So, I created a fake order (like the one in the course documentation at https://supersimplebackend.dev/documentation#post-orders) and use that for the rest of the project.